Move Main DB Object to App Instead of Farmer
