+++ tests/test_Farmer.py
@@ -1,6 +1,7 @@
 import unittest
+from dataserv.app import db
 from dataserv.Farmer import sha256
-from dataserv.Farmer import Farmer, db
+from dataserv.Farmer import Farmer
 
 
 class FarmerTest(unittest.TestCase):
